[rest/client]: move simple senders out of route utils and support statusCode override
 problem: simple send helpers don't use formatters, so don't need to specify type
          HTTP error codes can only be set via Restify errors
solution: move simple send helpers to new file
          allow statusCode to be specified
